Nasty hack to suppress an uninit warning.
"no warnings 'uninitialized'" would have suppressed a whole bunch of warnings
here and I didn't want that. However, I didn't want the warning to go
unchallenged, nor is a simple test for "&& defined $no" sufficient because we
want to guarantee that this path is taken (I could do "defined $no", but the
logic is getting complicated enough already).

Instead, we set $no to a known impossible value and cross our fingers. It
appears to work.